testsprite_tests/TC013_Admin_can_add_a_user_and_see_it_in_the_user_list.py
import asyncio
from playwright import async_api
from playwright.async_api import expect

async def run_test():
    pw = None
    browser = None
    context = None

    try:
        # Start a Playwright session in asynchronous mode
        pw = await async_api.async_playwright().start()

        # Launch a Chromium browser in headless mode with custom arguments
        browser = await pw.chromium.launch(
            headless=True,
            args=[
                "--window-size=1280,720",         # Set the browser window size
                "--disable-dev-shm-usage",        # Avoid using /dev/shm which can cause issues in containers
                "--ipc=host",                     # Use host-level IPC for better stability
                "--single-process"                # Run the browser in a single process mode
            ],
        )

        # Create a new browser context (like an incognito window)
        context = await browser.new_context()
        context.set_default_timeout(5000)

        # Open a new page in the browser context
        page = await context.new_page()

        # Interact with the page elements to simulate user flow
        # -> Navigate to http://localhost:5173
        await page.goto("http://localhost:5173")
        
        # -> Navigate to /login to load the login form so we can sign in as the Admin user.
        await page.goto("http://localhost:5173/login")
        
        # -> Click the 'Administrator' role button to open the Admin login form.
        frame = context.pages[-1]
        # Click element
        elem = frame.locator('xpath=/html/body/div/div/div[2]/div[3]/div/div[2]/button[4]').nth(0)
        await asyncio.sleep(3); await elem.click()
        
        # -> Fill the admin email into the email field (index 815), fill the password (index 821), then click Sign In (index 827).
        frame = context.pages[-1]
        # Input text
        elem = frame.locator('xpath=/html/body/div/div/div[2]/div[3]/div/div[2]/div/input').nth(0)
        await asyncio.sleep(3); await elem.fill('admin@eyeassist.demo')
        
        frame = context.pages[-1]
        # Input text
        elem = frame.locator('xpath=/html/body/div/div/div[2]/div[3]/div/div[2]/div[2]/input').nth(0)
        await asyncio.sleep(3); await elem.fill('Demo@1234')
        
        frame = context.pages[-1]
        # Click element
        elem = frame.locator('xpath=/html/body/div/div/div[2]/div[3]/div/div[2]/button').nth(0)
        await asyncio.sleep(3); await elem.click()
        
        # -> Open the User Management section by clicking the 'User Management' button (index 894).
        frame = context.pages[-1]
        # Click element
        elem = frame.locator('xpath=/html/body/div/div/div[2]/aside/button[2]').nth(0)
        await asyncio.sleep(3); await elem.click()
        
        # -> Click the 'User Management' button (index 894) to open the User Management page and reveal the Add User controls.
        frame = context.pages[-1]
        # Click element
        elem = frame.locator('xpath=/html/body/div/div/div[2]/aside/button[2]').nth(0)
        await asyncio.sleep(3); await elem.click()
        
        # -> Click the 'Add User' button to open the Add User form so we can fill required fields.
        frame = context.pages[-1]
        # Click element
        elem = frame.locator('xpath=/html/body/div/div/div[2]/main/div/div/div/div/button').nth(0)
        await asyncio.sleep(3); await elem.click()
        
        # -> Select the Role dropdown and choose 'Patient' (index 1293) before filling the rest of the form.
        frame = context.pages[-1]
        # Input text
        elem = frame.locator('xpath=/html/body/div/div/div[2]/main/div/div/div/div[2]/div/div[2]/form/div/div/input').nth(0)
        await asyncio.sleep(3); await elem.fill('Test Patient')
        
        frame = context.pages[-1]
        # Input text
        elem = frame.locator('xpath=/html/body/div/div/div[2]/main/div/div/div/div[2]/div/div[2]/form/div/div[2]/input').nth(0)
        await asyncio.sleep(3); await elem.fill('newpatient1@eyeassist.demo')
        
        # -> Click the 'Create User' button (index 1307) to submit the new user form, then verify the new user appears in the registered users list.
        frame = context.pages[-1]
        # Click element
        elem = frame.locator('xpath=/html/body/div/div/div[2]/main/div/div/div/div[2]/div/div[2]/form/div[2]/button').nth(0)
        await asyncio.sleep(3); await elem.click()
        
        # --> Test passed — verified by AI agent
        frame = context.pages[-1]
        current_url = await frame.evaluate("() => window.location.href")
        assert current_url is not None, "Test completed successfully"
        await asyncio.sleep(5)

    finally:
        if context:
            await context.close()
        if browser:
            await browser.close()
        if pw:
            await pw.stop()

asyncio.run(run_test())
    