requirements.txt
# ======================================================
# EYE-ASSISST — Backend Requirements
# ======================================================
# Python 3.10 – 3.11 recommended
# Install with:  pip install -r requirements.txt

# ── Web Framework ──────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ── Deep Learning ──────────────────────────────────────
# CPU-only build — swap for GPU build if CUDA is available
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

# ── Model Architecture ─────────────────────────────────
timm>=0.9.16

# ── Scientific / ML ────────────────────────────────────
numpy>=1.26.4
pandas>=2.2.0
scikit-learn>=1.4.0

# ── Image Processing ───────────────────────────────────
Pillow>=10.3.0
opencv-python>=4.9.0

# ── Explainability (Grad-CAM) ─────────────────────────
grad-cam>=1.5.0

# ── Visualisation ──────────────────────────────────────
matplotlib>=3.8.0
seaborn>=0.13.0

# ── Utilities ──────────────────────────────────────────
tqdm>=4.66.0
pyyaml>=6.0.1
rich>=13.7.0

# ── Dev / Notebooks (optional — remove for production) ─
# jupyter>=1.0.0
# ipykernel>=6.29.0
